Fixed try parse decimal point

+++ MQTTAPI/Model/Measurements.cs
@@ -3,6 +3,6 @@ namespace MQTTAPI.Model;
 public class Measurements
 {
     public decimal Temperature { get; set; }
-    public double Humidity { get; set; }
+    public decimal Humidity { get; set; }
     public DateTime Timestamp { get; set; }
 }
